+++ app/components/common/control/dropdown/types.ts
@@ -1,0 +1,4 @@
+export interface DropdownOption {
+  label: string;
+  data: any;
+}