+++ app/components/common/control/info-button/types.ts
@@ -1,0 +1,5 @@
+export interface Info {
+  title: string;
+  text: string;
+  variant?: ControlVariant;
+}