+++ app/routes/authenticated/settings/settings.route.ts
@@ -1,7 +1,7 @@
 import { action } from '@ember/object';
 import Route from '@ember/routing/route';
 import { service } from '@ember/service';
-import { DropdownOption } from 'potber-client/components/common/control/dropdown';
+import { DropdownOption } from 'potber-client/components/common/control/dropdown/types';
 import Session from 'potber-client/models/session';
 import RendererService from 'potber-client/services/renderer';
 import SessionService from 'potber-client/services/session';